refactor: cell.tsx 정리 - 바이탈 표시 커스텀 훅
